Refacto of safeFetch
