+++ src/api/helpers.ts
@@ -1,10 +1,10 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
 import { t } from '@lingui/macro'
-import jwtDecode from 'jwt-decode'
 
 import { navigationRef } from 'features/navigation/navigationRef'
 import { Headers, FailedToRefreshAccessTokenError } from 'libs/fetch'
 import { _ } from 'libs/i18n'
+import { decodeAccessToken } from 'libs/jwt'
 import { clearRefreshToken, getRefreshToken } from 'libs/keychain'
 import { storage } from 'libs/storage'
 
@@ -23,19 +23,6 @@ export async function getAuthenticationHeaders(options?: RequestInit): Promise<H
   return {}
 }
 
-interface AccessToken {
-  exp: number
-  fresh: false
-  iat: number
-  identity: string
-  jti: string
-  nbf: number
-  type: string
-  user_claims: {
-    user_id: number
-  }
-}
-
 // HOT FIX waiting for a better strategy
 const NotAuthenticatedCalls = [
   'native/v1/account',
@@ -47,14 +34,6 @@ const NotAuthenticatedCalls = [
   'native/v1/validate_email',
   'native/v1/offer',
 ]
-
-export const decodeAccessToken = (token: string) => {
-  try {
-    return jwtDecode<AccessToken>(token)
-  } catch {
-    return null
-  }
-}
 
 /**
  * For each http calls to the api, retrieves the access token and fetchs.
@@ -78,18 +57,15 @@ export const safeFetch = async (
   }
 
   // @ts-ignore
-  const authorizationHeader = options.headers?.['Authorization'] || ''
-  const token = authorizationHeader?.replace('Bearer ', '')
+  const authorizationHeader: string = options.headers?.['Authorization'] || ''
+  const token = authorizationHeader.replace('Bearer ', '')
+  const tokenContent = decodeAccessToken(token)
 
-  try {
-    // TODO: do not call jwtDecode when the token is undefined - modify jest setup so that jwtDecode is however called and returns a value
-    const tokenContent = jwtDecode<AccessToken>(token)
+  if (!tokenContent) {
+    return Promise.reject(navigateToLogin())
+  }
 
-    if (tokenContent.exp * 1000 <= new Date().getTime()) {
-      // TODO: do not throw an error
-      throw new Error('Token expired')
-    }
-  } catch (error) {
+  if (tokenContent.exp * 1000 <= new Date().getTime()) {
     try {
       const newAccessToken = await refreshAccessToken(api)
